finished DB connection and started login logic

+++ RentCars/RentCars/Model/Car.cs
@@ -2,7 +2,7 @@
 {
     public class Car
     {
-        public long CID { get; set; }
+        public int CID { get; set; }
         public string CarName { get; set; }
         public string CarBrand { get; set; }
         public int Seats { get; set; }
@@ -10,7 +10,7 @@
         public bool IsAutomatic { get; set; }
         public double HorsePwr { get; set; }
 
-        public Car(long id, string name, string brand)
+        public Car(int id, string name, string brand)
         {
             CID = id;
             CarName = name;
